[Tests] Coverage for TemplateResponse
